requirements.txt
torch
transformers
peft
torchvision
Pillow
pandas
tensorboard
pyarrow
fastparquet
scikit-learn